% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{add}
\alias{add}
\title{Add Two Objects}
\usage{
add(x, y)
}
\arguments{
\item{x}{An object.}

\item{y}{An object.}
}
\value{
The result of adding the two objects.
}
\description{
This function adds two objects. If both objects are numeric vectors, it performs element-wise addition. If one or both objects are strings, it concatenates them. For other objects, it attempts to use the \code{+} method defined for the class of the objects.
}
\keyword{math}
