% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research.R
\name{archiveCleanedPackage}
\alias{archiveCleanedPackage}
\title{Archive and clean Package Directory}
\usage{
archiveCleanedPackage(
  packageDir,
  output,
  removeHidden = TRUE,
  removeGitignore = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{packageDir}{A string specifying the path to the package directory.}

\item{output}{A string specifying the name of the output cleaned directory.
If the string ends with .zip, the function will create a zipped archive of
the cleaned directory.}

\item{removeHidden}{Logical; if \code{TRUE}, removes hidden files and folders
(starting with a dot). Default is \code{TRUE}.}

\item{removeGitignore}{Logical; if \code{TRUE}, removes files listed in the
\code{.gitignore}. Default is \code{TRUE}.}

\item{verbose}{Logical; if \code{TRUE}, prints additional information about the
cleaning process. Default is \code{FALSE}.}
}
\value{
A cleaned directory or a zipped file of the package directory.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function processes an R package directory, removing irrelevant files
(e.g., \code{.git}, \code{.gitignore}-listed files, etc.) and strips parts of lines
matching a specified regex from source code. It then creates a clean folder
or a zipped archive of the package without modifying the original directory.

MAJOR NOTE this is no where near perfect and the regexes used at the moment
for removing comments are prone to issues such as not detecting whether
the symbols are inside special strings.
}
\details{
The function cleans the package directory by:
\itemize{
\item Removing hidden files and folders like \code{.git}.
\item Excluding files listed in \code{.gitignore}.
\item Stripping out parts of lines matching a regex pattern from source files.
}

The output will be either a cleaned directory or a zipped archive of it.
}
\examples{
\dontrun{
# Clean and archive the package directory
archiveCleanedPackage("/path/to/package", "clean_package.zip")

# Clean package directory without zipping it
archiveCleanedPackage("/path/to/package", NULL)
}
}
\keyword{research}
