% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{bound}
\alias{bound}
\title{Bound a Number within a Range}
\usage{
bound(num, lowerbound, upperbound)
}
\arguments{
\item{num}{A numeric vector to be bounded.}

\item{lowerbound}{The lower bound of the range.}

\item{upperbound}{The upper bound of the range.}
}
\value{
A numeric vector with elements bounded within the specified range.
}
\description{
This function bounds a number within a specified range.
}
\examples{
bound(1, 0, 2)
bound(1:10, -1, 5)

}
\keyword{math}
