% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling.R
\name{closestWord}
\alias{closestWord}
\title{Find the Closest Word in a Set to a Given Word}
\usage{
closestWord(s, strset, distFunc = utils::adist)
}
\arguments{
\item{s}{A character string.}

\item{strset}{A set of character strings.}

\item{distFunc}{A function to compute distance between strings. Default is \code{utils::adist}.}
}
\value{
The closest word in the set to the given word.
}
\description{
This function finds the closest word in a set of words to a given word based on a specified distance function.
}
\examples{
# Find the closest word to "hello" in the set c("hallo", "hullo", "hey")
closestWord("hello", c("hallo", "hullo", "hey"))
}
\keyword{spelling}
