% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{colToRownames}
\alias{colToRownames}
\title{Convert a Column to Row Names}
\usage{
colToRownames(df, col, .remove = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{col}{A character string specifying the name of the column to convert to row names.}

\item{.remove}{A logical indicating whether to remove the selected column. Default is TRUE.}
}
\value{
A data frame with the specified column as row names.
}
\description{
This function converts a specified column of a data frame to row names.
}
\examples{
# Convert the 'ID' column to row names
df <- data.frame(ID = c("A", "B", "C"), Value = c(10, 20, 30))
colToRownames(df, "ID")
# Convert the 'ID' column to row names and keep the column
df <- data.frame(ID = c("A", "B", "C"), Value = c(10, 20, 30))
colToRownames(df, "ID", .remove = FALSE)
}
\seealso{
\code{\link[=mutateToRownames]{mutateToRownames()}}
}
\keyword{wrangling}
