% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higherOrderFunctions.R
\name{createMutator}
\alias{createMutator}
\title{Create a Mutator Function}
\usage{
createMutator(binaryOperator)
}
\arguments{
\item{binaryOperator}{A binary operator function to apply for the mutation.}
}
\value{
A function that takes a variable and a value, applying the binary operator to update the variable in the parent frame.
}
\description{
This function creates a mutator function based on a specified binary operator. The mutator function updates a variable in the parent frame by applying the binary operator with a given value.
}
\keyword{higherOrderFunctions}
