% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{divide}
\alias{divide}
\title{Divide Two Numbers}
\usage{
divide(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A numeric vector representing the quotient of the input vectors.
}
\description{
This function divides one number by another.
}
\examples{
# Divide two numeric vectors
divide(c(10, 20, 30), c(2, 4, 5))
}
\keyword{math}
