% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{enclose}
\alias{enclose}
\title{Enclose String with Specified Characters}
\usage{
enclose(x, left, right)
}
\arguments{
\item{x}{A character string to enclose.}

\item{left}{A character string to prepend.}

\item{right}{A character string to append.}
}
\value{
A new character string with \code{x} enclosed by \code{left} and \code{right}.
}
\description{
This function encloses a string with specified characters on the left and the right.
}
\examples{
enclose("text", "[", "]") # returns "[text]"
}
\keyword{character}
