% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{encloseBr}
\alias{encloseBr}
\title{Enclose String with Brackets}
\usage{
encloseBr(x)
}
\arguments{
\item{x}{A character string to enclose.}
}
\value{
A new character string with \code{x} enclosed by parentheses.
}
\description{
This function encloses a string with parentheses.
}
\examples{
encloseBr("text") # returns "(text)"
}
\keyword{character}
