% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{enumerate}
\alias{enumerate}
\title{Enumerate Elements with Indices}
\usage{
enumerate(..., zero_indexed = FALSE)
}
\arguments{
\item{...}{Vectors or lists to be enumerated.}

\item{zero_indexed}{A logical indicating whether indexing should start from zero. Default is FALSE.}
}
\value{
A list of lists, where each inner list contains an index and the corresponding elements from the input vectors or lists.
}
\description{
This function pairs elements of vectors or lists with their indices.
}
\keyword{iteration}
