% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{getAvgHex}
\alias{getAvgHex}
\title{Compute the Average of Hex Colors}
\source{
\url{https://stackoverflow.com/questions/649454}
}
\usage{
getAvgHex(...)
}
\arguments{
\item{...}{Hex color values as character strings.}
}
\value{
A single hex color value representing the average of the input colors.
}
\description{
This function computes the average color of the provided hex color values.
}
\keyword{color}
