% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling.R
\name{getChar}
\alias{getChar}
\title{Get a Character at a Specific Index}
\usage{
getChar(x, index)
}
\arguments{
\item{x}{A character string.}

\item{index}{The index of the character to retrieve.}
}
\value{
The character at the specified index.
}
\description{
This function retrieves a character at a specific index from a string.
}
\examples{
# Get the character at index 2
getChar("hello", 2)
}
\keyword{spelling}
