% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDevelopment.R
\name{getExistingFilePath}
\alias{getExistingFilePath}
\title{Get Existing File Path}
\usage{
getExistingFilePath(filePath, dir = ".")
}
\arguments{
\item{filePath}{The relative or absolute path to the file.}

\item{dir}{The directory in which to check for the file, default is the current directory ('.').}
}
\value{
The full path to the file if it exists.
}
\description{
This function checks whether a specified file exists within a given directory.
It is a simple wrapper that helps in error handling by throwing an error if the
directory does not exist or if the file is not found within that directory.
This ensures that only valid file paths are returned.
}
\examples{
\donttest{
getExistingFilePath("DESCRIPTION")
}
}
\keyword{packageDevelopment}
