% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{getfirst}
\alias{getfirst}
\title{Get the First Elements of a Vector or List}
\usage{
getfirst(x, n = 1)
}
\arguments{
\item{x}{A vector, list, or other supported data type.}

\item{n}{An integer specifying the number of elements to retrieve from the start. Default is 1.}
}
\value{
The first \code{n} elements of the input.
}
\description{
This function retrieves the first \code{n} elements of a vector or list.
}
\keyword{indexing}
