% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{getlast}
\alias{getlast}
\title{Get the Last Elements of a Vector or List}
\usage{
getlast(x, n = 1)
}
\arguments{
\item{x}{A vector, list, or other supported data type.}

\item{n}{An integer specifying the number of elements to retrieve from the end. Default is 1.}
}
\value{
The last \code{n} elements of the input.
}
\description{
This function retrieves the last \code{n} elements of a vector or list.
}
\keyword{indexing}
