% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDevelopment.R
\name{getPkgKeywords}
\alias{getPkgKeywords}
\title{Get Keywords from R Package Documentation}
\usage{
getPkgKeywords(pkg = ".", asDistribution = FALSE)
}
\arguments{
\item{pkg}{The path to the R package directory.}

\item{asDistribution}{Logical; if FALSE, returns a character vector of unique keywords. If TRUE, returns a table with the frequency of each keyword.}
}
\value{
If \code{asDistribution} is FALSE, a character vector of unique keywords is returned.
If \code{asDistribution} is TRUE, a table of keywords and their frequencies is returned.
}
\description{
This function retrieves keywords from all package documentation files located
in the \verb{/man} directory of the specified R package. It can return a unique list
of keywords or a frequency distribution of these keywords as a \code{table} object.
}
\examples{
getPkgKeywords()
getPkgKeywords(asDistribution = TRUE)

}
\keyword{packageDevelopment}
