% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getUniquePairsUpTo}
\alias{getUniquePairsUpTo}
\title{Generate Unique Pairs Up To a Number}
\usage{
getUniquePairsUpTo(x, oneIndexed = TRUE)
}
\arguments{
\item{x}{An integer specifying the upper limit for pairs.}

\item{oneIndexed}{A logical indicating whether the pairs should be one
indexed. Default is TRUE.}
}
\value{
A list of unique pairs of integers up to the specified number.
}
\description{
This function generates all unique pairs of integers up to a given number.
}
\examples{
# Generate unique pairs up to 3 (one-indexed)
getUniquePairsUpTo(3)
# Generate unique pairs up to 3 (zero-indexed)
getUniquePairsUpTo(3, oneIndexed = FALSE)
}
\keyword{iteration}
