% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{ind}
\alias{ind}
\title{Get Index from Enumerated Element}
\usage{
ind(elem)
}
\arguments{
\item{elem}{An enumerated element.}
}
\value{
The index of the enumerated element.
}
\description{
This function extracts the index from an enumerated element.
}
\examples{
# Extract index from an enumerated element
elem <- list(1, "a")
ind(elem)
}
\seealso{
\code{\link[=enumerateit]{enumerateit()}}
}
\keyword{iteration}
