% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInitialization.R
\name{initList}
\alias{initList}
\title{Initialize a List}
\usage{
initList(x, initVal = NULL)
}
\arguments{
\item{x}{A character vector as names, or an numeric indicating the size of the list.}

\item{initVal}{an optional initial value for all elements of the list.}
}
\value{
A list of the specified size and names, optionally initialized with a value.
}
\description{
This function initializes a list based on size or names, with an optional initial value.
}
\keyword{dataInitialization}
