% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{initTestthat}
\alias{initTestthat}
\title{Initialize Testthat Files}
\usage{
initTestthat(
  rDir = "R",
  testDir = "tests/testthat",
  ignore = c("-package.R$", "-class.R$", "^data.R$", "^zzz.R$", "^RcppExports.R$")
)
}
\arguments{
\item{rDir}{The directory containing R source files. Default is "R".}

\item{testDir}{The directory where \code{testthat} files should be created. Default is "tests/testthat".}

\item{ignore}{A character vector specifying regex patterns of files to ignore.}
}
\value{
No return value, called for side effects.
}
\description{
This function scans all files in the specified R directory, excluding some based on the patterns provided in the \code{ignore} argument, and creates \code{testthat} files if they are missing.
}
\examples{
# Initialize testthat files in the default directories
initTestthat()
# Initialize testthat files in a custom R directory and test directory
initTestthat(rDir = "src", testDir = "tests")
# Initialize testthat files, ignoring additional patterns
initTestthat(ignore = c("-package.R$", "-class.R$", "^data.R$", "^zzz.R$", "^RcppExports.R$", "helper-.*\\\\.R$"))
}
\keyword{testing}
