% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInitialization.R
\name{initV}
\alias{initV}
\title{Initialize a Vector}
\usage{
initV(typeFunc, x, initVal = NULL)
}
\arguments{
\item{typeFunc}{A character string indicating the type of the vector or a function to create the vector.}

\item{x}{The length of the vector.}

\item{initVal}{An optional initial value to fill the vector.}
}
\value{
A vector of the specified type and size, optionally initialized with a value.
}
\description{
This function initializes a vector based on a specified type and size, with an optional initial value.
}
\examples{
# Create a numeric vector of length 5
initV("numeric", 5)
# Create a logical vector of length 3 initialized with TRUE
initV("logical", 3, TRUE)
}
\keyword{dataInitialization}
