% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageLoading.R
\name{installAndLoad}
\alias{installAndLoad}
\title{Install and Load Packages from Various Sources}
\usage{
installAndLoad(
  cran = NULL,
  bioc = NULL,
  gh = NULL,
  lib = .libPaths()[1],
  verbose = FALSE
)
}
\arguments{
\item{cran}{A character vector of CRAN package names to install and load.}

\item{bioc}{A character vector of Bioconductor package names to install and load.}

\item{gh}{A character vector of GitHub repositories in the format "username/repo" to install and load.}

\item{lib}{a character vector of the directory of the library to install the packages in.}

\item{verbose}{a logical indicating whether to display additional verbal cues}
}
\value{
None, function is used for side effects (installation and loading of packages).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function is designed to facilitate the installation and loading of R packages
from CRAN, Bioconductor, or GitHub. It checks if packages are installed; if not, it installs them,
and then loads them into the R session. Most importantly, it can handle an arbitrary number
of packages at once and install all possible dependencies.
}
\details{
The function first checks if necessary namespaces (\code{BiocManager} for Bioconductor packages and
\code{devtools} for GitHub packages) are installed, and installs them if they are not present.
It then proceeds to install and load packages from the specified sources. For GitHub packages,
installation is done using \code{\link[devtools:remote-reexports]{devtools::install_github()}}.
}
\examples{
\dontrun{
# Install and load CRAN packages
installAndLoad(cran = c("ggplot2", "data.table"))

# Install and load a Bioconductor package
installAndLoad(bioc = c("GenomicFeatures"))

# Install and load a package from GitHub
installAndLoad(gh = c("hadley/lubridate"))

# Install and load packages from mixed sources
installAndLoad(cran = c("dplyr"), bioc = c("BiocGenerics"), gh = c("r-lib/testthat"))
}

}
\keyword{packageLoading}
