% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling.R
\name{isCamelCase}
\alias{isCamelCase}
\title{Check if String is camelCase}
\usage{
isCamelCase(x)
}
\arguments{
\item{x}{A character string to check.}
}
\value{
TRUE if the string is camelCase, FALSE otherwise.
}
\description{
This function checks if a given string adheres to camelCase naming
conventions, starting with a lowercase letter followed by any
combination of upper and lower case letters.
}
\examples{
isCamelCase("camelCase")   # returns TRUE
isCamelCase("CamelCase")   # returns FALSE
isCamelCase("camelcase")   # returns TRUE

}
\keyword{spelling}
