% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{isEven}
\alias{isEven}
\title{Check if a Number is Even}
\usage{
isEven(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A logical vector indicating whether each element is even.
}
\description{
This function checks if a number is even.
}
\examples{
# Check if numbers are even
isEven(c(1, 2, 3, 4))
}
\keyword{math}
