% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{isOdd}
\alias{isOdd}
\title{Check if a Number is Odd}
\usage{
isOdd(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A logical vector indicating whether each element is odd.
}
\description{
This function checks if a number is odd.
}
\keyword{math}
