% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling.R
\name{isPascalCase}
\alias{isPascalCase}
\title{Check if String is PascalCase}
\usage{
isPascalCase(x)
}
\arguments{
\item{x}{A character string to check.}
}
\value{
TRUE if the string is PascalCase, FALSE otherwise.
}
\description{
This function checks if a given string adheres to PascalCase naming conventions,
starting with an uppercase letter followed by any combination of upper and lower case letters.
}
\examples{
isPascalCase("PascalCase") # returns TRUE
isPascalCase("pascalCase") # returns FALSE
isPascalCase("Pascalcase") # returns FALSE
}
\keyword{spelling}
