% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling.R
\name{isVowel}
\alias{isVowel}
\title{Check if a Character is a Vowel}
\usage{
isVowel(x)
}
\arguments{
\item{x}{A character.}
}
\value{
TRUE if the character is a vowel, FALSE otherwise.
}
\description{
This function checks if a character is a vowel.
}
\examples{
# Check if 'a' is a vowel
isVowel("a")
# Check if 'b' is a vowel
isVowel("b")
}
\keyword{spelling}
