% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{joinRegex}
\alias{joinRegex}
\title{Join regex expressions by union}
\usage{
joinRegex(...)
}
\arguments{
\item{...}{character vectors of the regex expressions to join. Both
vectors and individual characters of any length will work}
}
\value{
a character of the unioned regex
}
\description{
This function simply joins a vector of regex characters by union,
and produces a single character regex in the form of \code{(foo)|(bar)}.
}
\examples{
joinRegex(c("^foo", "bar$"))
joinRegex("^foo", "bar$", "[bB]az")

}
\keyword{regex}
