% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInitialization.R
\name{namedNumericToTable}
\alias{namedNumericToTable}
\title{Convert Named Numeric Vector to Table}
\usage{
namedNumericToTable(x)
}
\arguments{
\item{x}{A named numeric vector.}
}
\value{
A table with the same names and values as the input vector.
}
\description{
This function converts a named numeric vector to a table.
}
\examples{
# Convert a named numeric vector to a table
vec <- c(a = 1, b = 2, c = 3)
namedNumericToTable(vec)
}
\keyword{dataInitialization}
