% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling.R
\name{prependIndefArticle}
\alias{prependIndefArticle}
\title{Prepend an Indefinite Article to a String}
\usage{
prependIndefArticle(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
The string with an indefinite article prepended.
}
\description{
This function prepends an indefinite article ("a" or "an") to a string based on whether it starts with a vowel or not.
}
\examples{
# Prepend an indefinite article to "apple"
prependIndefArticle("apple")
# Prepend an indefinite article to "banana"
prependIndefArticle("banana")
}
\keyword{spelling}
