% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDevelopment.R
\name{removeVdiffrNewSnapShots}
\alias{removeVdiffrNewSnapShots}
\alias{rmns}
\title{Remove New Snapshots from vdiffr Tests}
\usage{
removeVdiffrNewSnapShots(
  pkg = ".",
  snapDir = file.path("tests", "testthat", "_snaps")
)

rmns(pkg = ".", snapDir = file.path("tests", "testthat", "_snaps"))
}
\arguments{
\item{pkg}{The path to the package directory.}

\item{snapDir}{The path to the directory containing the snapshots. Default
is \verb{tests/testthat/_snaps}. If this directory isn't valid, nothing happens.}
}
\value{
NULL (invisible) - used for side effects
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function removes new snapshots created by \code{vdiffr} tests. It is useful
when you want to remove new snapshots that were created during testing and
are no longer needed.
}
\examples{
/donttest{
removeVdiffrNewSnapShots()
}

}
\keyword{testing}
