% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{rowDiff}
\alias{rowDiff}
\title{Returns rows that are different between two data frames}
\usage{
rowDiff(df1, df2)
}
\arguments{
\item{df1}{A data frame.}

\item{df2}{A data frame.}
}
\value{
A data frame containing the rows that are different between
the two data frames. Note that rownames are removed. If there
are no differences, the function returns a
row-less data frame with the same column names.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function compares two data frames and returns the rows that are
different between them. The function assumes that the data frames have
the same column names and types and errors/has undefined behaviour
otherwise.
}
\examples{
rowDiff(mtcars[1:10, ], mtcars[5:15, ])
}
\keyword{wrangling}
