% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scaleHex}
\alias{scaleHex}
\title{Scale the Brightness of a Hex Color}
\usage{
scaleHex(hex, scaleFactor)
}
\arguments{
\item{hex}{Hex color values as characters.}

\item{scaleFactor}{A numeric value to scale the brightness.
A value of 1 returns the original color.}
}
\value{
A hex color value with adjusted brightness.
}
\description{
This function scales the brightness of hex colors by a given factor.
}
\examples{
scaleHex("#404040", 2)

}
\keyword{color}
