% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{setAt}
\alias{setAt}
\title{Set a value at a specified index in an object}
\usage{
setAt(x, i, value)
}
\arguments{
\item{x}{An object to modify.}

\item{i}{The index at which the value should be set.}

\item{value}{The value to assign at the specified index.}
}
\value{
The modified object, with the value set at the specified index.
}
\description{
Chainable functional alias of \code{x[i] <- value}.
}
\examples{
x <- c(1, 2, 3, 4, 5)
setAt(x, 2, 10)
}
\keyword{indexing}
