% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{setCol}
\alias{setCol}
\title{Set values in a column of an object}
\usage{
setCol(x, j, value)
}
\arguments{
\item{x}{An object that supports the \code{"[<-"} method, such as a matrix or
data frame.}

\item{j}{Column index}

\item{value}{The value to assign to the specified index/indices.}
}
\value{
The modified object.
}
\description{
Chainable functional alias of \code{x[, j] <- value}.
}
\examples{
mat <- matrix(1:9, nrow = 3)
setCol(mat, j = 2, value = 5)  # Set entire second column to 5
}
\seealso{
\code{\link[=setIndex]{setIndex()}}
}
\keyword{indexing}
