% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{setColnames}
\alias{setColnames}
\title{Set Column Names}
\usage{
setColnames(object, newColnames)
}
\arguments{
\item{object}{A data frame or matrix.}

\item{newColnames}{A character vector specifying the new column names.}
}
\value{
The data frame or matrix with updated column names.
}
\description{
This function sets new column names for a given data frame or matrix.
}
\examples{
# Set new column names for a data frame
df <- data.frame(A = c(1, 2, 3), B = c(4, 5, 6))
setColnames(df, c("X", "Y"))
}
\keyword{wrangling}
