% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{setIndex}
\alias{setIndex}
\title{Set values in an object using indexing}
\usage{
setIndex(x, i, j, value)
}
\arguments{
\item{x}{An object that supports the \code{"[<-"} method, such as a matrix or
data frame.}

\item{i}{Row index}

\item{j}{Column index}

\item{value}{The value to assign to the specified index/indices.}
}
\value{
The modified object.
}
\description{
Chainable functional alias for \code{x[i, j] <- value}. It provides flexibility
for setting values using either one or two indices. Note that if you only
want to modify columns (\code{j}), the argument along with the \code{value} must be
named. Alternatively use \link{setCol}. if neither \code{i} nor \code{j} is provided, an
error is thrown.
}
\examples{
mat <- matrix(1:9, nrow = 3)
setIndex(mat, 1, 2, 10)  # Set value at position (1, 2)
setIndex(mat, 1, value = 20)  # Set entire first row to 20
setIndex(mat, j = 3, value = 30)  # Set entire third column to 30
}
\seealso{
\code{\link[=setCol]{setCol()}}, \code{\link[=setRow]{setRow()}}
}
\keyword{indexing}
