% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{setRow}
\alias{setRow}
\title{Set values in a row of an object}
\usage{
setRow(x, i, value)
}
\arguments{
\item{x}{An object that supports the \code{"[<-"} method, such as a matrix or
data frame.}

\item{i}{Row index}

\item{value}{The value to assign to the specified index/indices.}
}
\value{
The modified object.
}
\description{
Chainable functional alias of \code{x[i, ] <- value}.
}
\examples{
mat <- matrix(1:9, nrow = 3)
setRow(mat, i = 1, value = 10)  # Set entire first row to 10
}
\seealso{
\code{\link[=setIndex]{setIndex()}}
}
\keyword{indexing}
