% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{setRownames}
\alias{setRownames}
\title{Set Row Names}
\usage{
setRownames(object, newRownames)
}
\arguments{
\item{object}{A data frame or matrix.}

\item{newRownames}{A character vector specifying the new row names.}
}
\value{
The data frame or matrix with updated row names.
}
\description{
This function sets new row names for a given data frame or matrix.
}
\examples{
# Set new row names for a data frame
df <- data.frame(A = c(1, 2, 3), B = c(4, 5, 6))
setRownames(df, c("row1", "row2", "row3"))
}
\keyword{wrangling}
