% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling.R
\name{splitSnake}
\alias{splitSnake}
\title{Split Snake Case String}
\usage{
splitSnake(x)
}
\arguments{
\item{x}{A character string in snake_case to be split.}
}
\value{
A list of character vectors, each containing the parts of the string split at underscores.
}
\description{
This function splits a string formatted in snake_case into its component words,
using underscores as delimiters. It is useful for parsing identifiers or variable
names that follow snake_case naming conventions.
}
\examples{
splitSnake("this_is_snake_case")
splitSnake("another_example_here")

}
\keyword{spelling}
