% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling.R
\name{startsWithVowel}
\alias{startsWithVowel}
\title{Check if a String Starts with a Vowel}
\usage{
startsWithVowel(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
TRUE if the string starts with a vowel, FALSE otherwise.
}
\description{
This function checks if a string starts with a vowel.
}
\keyword{spelling}
