% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling.R
\name{stripSpaces}
\alias{stripSpaces}
\title{Remove Spaces from a String}
\usage{
stripSpaces(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
The string with spaces removed.
}
\description{
This function removes spaces from a character string.
}
\keyword{spelling}
