% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{subtract}
\alias{subtract}
\title{Subtract Two Numbers}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A numeric vector representing the difference between the input vectors.
}
\description{
This function subtracts one number from another.
}
\keyword{math}
