% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDevelopment.R
\name{syncYMLKeywordRefs}
\alias{syncYMLKeywordRefs}
\alias{sykr}
\title{Add or Sync Keyword References in Pkgdown YAML File}
\usage{
syncYMLKeywordRefs(
  pkg = ".",
  ymlPath = "_pkgdown.yml",
  mode = "append",
  sectionName = "reference",
  modifier = trySplitVar
)

sykr(
  pkg = ".",
  ymlPath = "_pkgdown.yml",
  mode = "append",
  sectionName = "reference",
  modifier = trySplitVar
)
}
\arguments{
\item{pkg}{The path to the R package directory.}

\item{ymlPath}{The path to the \verb{_pkgdown.yml} file, relative or absolute.
Default is "_pkgdown.yml" within the package directory.}

\item{mode}{Character string, "append" or "overwrite", determining how keywords are handled.}

\item{sectionName}{The name of the YAML section to modify. Default is "reference".}

\item{modifier}{A function to modify the keyword list; defaults to \code{trySplitVar}.}
}
\description{
Updates or creates the \verb{_pkgdown.yml} file in a specified R package directory.
It appends or overwrites entries in a YAML section based on unique keywords extracted
from the package documentation, transformed using \code{trySplitVar} for more readable titles.
}
\keyword{packageDevelopment}
