% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInitialization.R
\name{tableToNumeric}
\alias{tableToNumeric}
\title{Convert a Table to Numeric}
\usage{
tableToNumeric(x)
}
\arguments{
\item{x}{A table to be converted.}
}
\value{
A numeric vector with names preserved from the table.
}
\description{
This function converts a table to a numeric vector.
}
\keyword{dataInitialization}
