% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_quietly_that}
\alias{test_quietly_that}
\title{Run a Testthat test Quietly}
\usage{
test_quietly_that(desc, code)
}
\arguments{
\item{desc}{A description of the test.}

\item{code}{The code to be tested.}
}
\value{
No return value, called for side effects.
}
\description{
This function runs a \code{test_that} block quietly, suppressing messages and
output from any verbose functions.
}
\examples{
# Run a test quietly
test_quietly_that("quiet test example", {
    testthat::expect_equal(1 + 1, 2)
})
}
\keyword{testing}
