% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling.R
\name{trySplitVar}
\alias{trySplitVar}
\title{Try to Split Variable Names Based on Naming Convention}
\usage{
trySplitVar(x)
}
\arguments{
\item{x}{A character string or vector to be analyzed and split.}
}
\value{
A list of character vectors, each containing the parts of the string
split according to its naming convention or the original string if no
convention matches.
}
\description{
This function attempts to split characters into its component words based on
camelCase, PascalCase, or snake_case conventions. If the string does not match
any of these conventions, it returns the original string.
}
\examples{
trySplitVar("camelCaseExample")
trySplitVar("PascalCaseExample")
trySplitVar("snake_case_example")
trySplitVar("no_special_case")
}
\seealso{
\code{\link{splitCamel}}, \code{\link{splitPascal}}, \code{\link{splitSnake}},
\code{\link{isCamelCase}}, \code{\link{isPascalCase}}, \code{\link{isSnakeCase}}
}
\keyword{spelling}
