% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spelling.R
\name{trySplitVarStr}
\alias{trySplitVarStr}
\title{Try to Split Variable Names Based on Naming Convention}
\usage{
trySplitVarStr(x, conseq = TRUE, strictSnake = FALSE)
}
\arguments{
\item{x}{A character string or vector to be analyzed and split.}

\item{conseq}{A logical indicating whether the \code{conseq} argument in \code{\link[=splitCamel]{splitCamel()}}/
\code{\link[=splitPascal]{splitPascal()}} should be \code{TRUE} or \code{FALSE}.}

\item{strictSnake}{A logical indicating the \code{strict} argument in \code{\link[=isSnakeCase]{isSnakeCase()}}.}
}
\value{
A list of character vectors, each containing the parts of the string
split according to its naming convention or the original string if no
convention matches.
}
\description{
This function attempts to split characters into its component words based on
camelCase, PascalCase, or snake_case conventions. If the string does not match
any of these conventions, it returns the original string.
}
\examples{
trySplitVarStr("camelCaseExample")
trySplitVarStr("PascalCaseExample")
trySplitVarStr("snake_case_example")
trySplitVarStr("some|random|case")

}
\seealso{
\code{\link{splitCamel}}, \code{\link{splitPascal}}, \code{\link{splitSnake}},
\code{\link{isCamelCase}}, \code{\link{isPascalCase}}, \code{\link{isSnakeCase}}
}
\keyword{spelling}
