% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{val1}
\alias{val1}
\title{Get First Value from Enumerated Element}
\usage{
val1(elem)
}
\arguments{
\item{elem}{An enumerated element.}
}
\value{
The first value in the enumerated element.
}
\description{
This function extracts the first value from an enumerated element.
}
\examples{
# Extract the first value from an enumerated element
elem <- list(1, "a", "b")
val1(elem)
}
\keyword{iteration}
