% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{zip}
\alias{zip}
\title{Zip Multiple Vectors or Lists}
\usage{
zip(...)
}
\arguments{
\item{...}{Vectors or lists to be combined.}
}
\value{
A list of lists, where each inner list contains the elements from the corresponding positions in the input vectors or lists.
}
\description{
This function combines multiple vectors or lists element-wise into a list of lists.
}
\examples{
# Zip two vectors
zip(c(1, 2, 3), c("a", "b", "c"))
# Zip three vectors
zip(c(1, 2), c("x", "y"), c(TRUE, FALSE))
}
\keyword{iteration}
