% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{zipit}
\alias{zipit}
\title{Zip Multiple Vectors or Lists}
\usage{
zipit(...)
}
\arguments{
\item{...}{Vectors or lists to be combined.}
}
\value{
A list of lists, where each inner list contains the elements from
the corresponding positions in the input vectors or lists.
}
\description{
This function combines multiple vectors or lists element-wise into
a list of lists. It's a slightly lighter weight alternative to
\code{itertools::izip()}
}
\examples{
# Zip two vectors
zipit(c(1, 2, 3), c("a", "b", "c"))
# Zip three vectors
zipit(c(1, 2), c("x", "y"), c(TRUE, FALSE))
}
\seealso{
\code{\link[=enumerateit]{enumerateit()}}
}
\keyword{iteration}
