% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{colToRownames}
\alias{colToRownames}
\title{Convert a Column to Row Names}
\usage{
colToRownames(df, colname, matrix = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{colname}{A character string specifying the name of the column to convert to row names.}

\item{matrix}{A logical indicating whether to return the result as a matrix. Default is FALSE.}
}
\value{
A data frame with the specified column as row names. If \code{matrix} is TRUE, a matrix is returned.
}
\description{
This function converts a specified column of a data frame to row names.
}
\examples{
# Convert the 'ID' column to row names
df <- data.frame(ID = c("A", "B", "C"), Value = c(10, 20, 30))
colToRownames(df, "ID")
# Convert the 'ID' column to row names and return as matrix
colToRownames(df, "ID", matrix = TRUE)
}
\keyword{wrangling}
