% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastUtils-package.R
\docType{package}
\name{FastUtils-package}
\alias{FastUtils}
\alias{FastUtils-package}
\title{FastUtils: Fast, readable utility functions}
\description{
for general use. All functions have vectorized implementations whenever possible.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/Qile0317/FastUtils/issues/}
}

}
\author{
\strong{Maintainer}: Qile Yang \email{qile.yang@berkeley.edu} [copyright holder]

}
\keyword{internal}
