% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastUtils-package.R
\docType{package}
\name{FastUtils-package}
\alias{FastUtils}
\alias{FastUtils-package}
\title{FastUtils: Fast, Readable Uutility Functions}
\description{
A wide variety of tools for general data analysis, wrangling, package development. All functions have vectorized implementations whenever possible.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Qile0317/FastUtils}
  \item \url{https://qile0317.github.io/FastUtils/}
  \item Report bugs at \url{https://github.com/Qile0317/FastUtils/issues/}
}

}
\author{
\strong{Maintainer}: Qile Yang \email{qile.yang@berkeley.edu} [copyright holder]

}
\keyword{internal}
