% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{getAvgHex}
\alias{getAvgHex}
\title{Compute the Average of Hex Colors}
\source{
\url{https://stackoverflow.com/questions/649454}
}
\usage{
getAvgHex(...)
}
\arguments{
\item{...}{Hex color values as character strings. Could also be any number
of character vectors (including lists) which will all be coerced into a single
characters, assuming they are valid hex codes.}
}
\value{
A single hex color character representing the average of the input colors.
}
\description{
This function computes the average color of the provided hex color values.
}
\examples{
getAvgHex("#00000", "#FF00FF")
getAvgHex(c("#008040", "#00000", "#FF00FF"))

# very nonstandard but possible way to input hexes. Essentially,
any combination of vectors will work.
getAvgHex(list("#008040", "#000000"), "#FF00FF", c("#FF00FF"))

}
\keyword{color}
