% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInitialization.R
\name{initDataFrameWithColnames}
\alias{initDataFrameWithColnames}
\title{Initialize a DataFrame with Column Names}
\usage{
initDataFrameWithColnames(colnames)
}
\arguments{
\item{colnames}{A character vector specifying the names of the columns for the data frame.}
}
\value{
A data frame with the specified column names.
}
\description{
This function creates an empty data frame and assigns the specified column names with
zero rows.
}
\examples{
# Create a data frame with specified column names initialized with NA
initDataFrameWithColnames(c("Name", "Age", "Gender"))

}
\keyword{dataInitialization}
