% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInitialization.R
\name{initEmptyTable}
\alias{initEmptyTable}
\title{Initialize an Empty Table}
\usage{
initEmptyTable()
}
\value{
An empty table structure.
}
\description{
This function initializes an empty table.
}
\examples{
# Create an empty table
initEmptyTable()
}
\keyword{dataInitialization}
