% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{isBound}
\alias{isBound}
\title{Check if a Number is within a Range}
\usage{
isBound(num, lowerbound, upperbound)
}
\arguments{
\item{num}{A numeric vector to be checked.}

\item{lowerbound}{The lower bound of the range.}

\item{upperbound}{The upper bound of the range.}
}
\value{
A logical vector indicating whether each element is within the specified range.
}
\description{
This function checks if a number is within a specified range.
}
\examples{
isBound(1, 0, 2)
isBound(1:10, -1, 5)

}
\keyword{math}
