% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{multiply}
\alias{multiply}
\title{Multiply Two Numbers}
\usage{
multiply(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A numeric vector representing the product of the input vectors.
}
\description{
This function multiplies two numbers.
}
\examples{
# Multiply two numeric vectors
multiply(c(2, 3, 4), c(5, 6, 7))
}
\keyword{math}
