% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{mutateToRownames}
\alias{mutateToRownames}
\title{Mutate columns to Row Names}
\usage{
mutateToRownames(.data, expr, .remove = FALSE, .uniqueSep = ".")
}
\arguments{
\item{.data}{A data frame.}

\item{expr}{A tidy evaluation expression specifying the columns to use for
the new row names.}

\item{.remove}{A logical indicating whether to remove the columns used in
\code{expr}. Default is FALSE. Note that columns are detected with \code{\link[=all.vars]{all.vars()}}.}

\item{.uniqueSep}{A character string to separate duplicate row names when
ensuring uniqueness
with \code{\link[=make.unique]{make.unique()}}. Default is ".".}
}
\value{
A data frame with updated row names.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function sets new row names for a data frame based on a tidy
evaluation expression.
}
\examples{
library(dplyr)

# if you just wish to make a column (`wt`) into a rowname:
mtcars \%>\%
    head() \%>\%
    mutateToRownames(wt)

# simple example of using tidy evaluation
mtcars \%>\%
    head() \%>\%
    mutateToRownames(wt + 3*vs)

# use .remove = TRUE to get rid of the column if unneeded
mtcars \%>\%
    mutateToRownames(qsec + (gear / carb) * wt, .remove = TRUE)

}
\keyword{wrangling}
