% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactivity.R
\name{quietly}
\alias{quietly}
\title{Suppress Messages and Output}
\usage{
quietly(e)
}
\arguments{
\item{e}{An expression to evaluate.}
}
\value{
The result of the expression with messages suppressed and output
captured.
}
\description{
This function suppresses messages and captures output from an expression.
Note that this does NOT return the result of the expression! It is
intended to be used in contexts where certain operations with side
effects are performed but the verbosity should be suppressed.
}
\examples{

quietly(print(1))

quietly({
    print(1)
    print(2)
    print(3)
})

a <- 1
quietly({
    a <- a + 1
    print(a)
})
print(a)
#> 2

}
\keyword{interactivity}
