% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character.R
\name{substrEnd}
\alias{substrEnd}
\title{Extract Substring from Start to End Difference}
\usage{
substrEnd(x, start, endDiff)
}
\arguments{
\item{x}{A character string from which the substring is extracted.}

\item{start}{The starting position for the substring extraction.}

\item{endDiff}{The difference to subtract from the string length to determine the end position.}
}
\value{
A substring of the input character string.
}
\description{
This function extracts a substring from a given start position to the position determined by subtracting \code{endDiff} from the string length.
}
\examples{
substrEnd("12345", 1, 1)
substrEnd("12345", 1, 2)
substrEnd("12345", 2, 3)

}
\seealso{
\code{\link[=substr]{substr()}}
}
\keyword{character}
